publisher/requirements.txt
google-cloud-pubsub>=2.16.0
faker>=19.6.1
